Fixes #124
Happy now Lauren?
