+++ testing/src/test/resources/csv/flw_language_update.csv
@@ -1,0 +1,3 @@
+NMS FLW-ID,MCTS FLW-ID,MSISDN,LANGUAGE CODE
+72185,210302604211400029,1000000000,hi
+,,2000000000,hi
